nft-metadata/bcdb-metadata-updater/requirements.txt
golem-base-sdk
python-dotenv
